GX2-16267 Renamed urlData and responseType files
