+++ cypress/support/types/GX2-16267-responseType.ts
